Rename "inspectShell" to "inspect_shell".
