+++ src/funfuzz/js/inspect_shell.py
